// src/generated/RoleBinding.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Id } from "./Id";
import type { Role } from "./Role";

export type RoleBinding = { role: Role, primary_binding_id: Id, 
/**
 * Ordered; user-configured; never inferred.
 */
fallback_binding_ids: Array<Id>, };